refactor : Mate 그룹화
Mate
   - MateMain
   - MateAllNight
